requirements.txt
gradio>=3.0
transformers>=4.0
Pillow>=9.0
